Ohifv3 (#1562)
* Add extension and mode for MONAI Label


* Add toolGroups to mode


* Extension and model working in class format


* Connection to MONAI Label server and label fetching are working now


* Add probe tool


* add segmentation load/rendering for OHIF v3 (#1525)

* working segmentation load

* segmentation panel

* add frame of reference uid to monai segs

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------


* Update segmentatio reader function


* Model inference and label rendering working


* Convert points from world to ijk - remove all clicks


* Increase caching files when connecting to DICOM server


* Add active learning tab


* Update style for active learning tab


* Hide the right panel


* Send centroids to viewer


* Send segmentation to DICOM Web


* Fix centroids - json compatible


* Update gitignore - add build script


* feat(ohifv3): parse segment centroids for jump and make smartEdit work (#1531)

* working segmentation load

* segmentation panel

* add frame of reference uid to monai segs

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* make probe tool grab color for selected segment

* parse centroids and add to segmentation

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------


* Fix indexes for centroids


* Add image


* Add segmentation panel


* DICOM-SEG writer is working


* DeepEdit working on OHIFv3


* Update tab name


* Update gitignore


* OHIFv3 working


* Update bash scrip to build OHIFv3


* COOP solved - using orthanc instead of proxy


* Update print statement


* Change tab background


* feat(segmentation tools): working new segmentation panel and brushes (#1545)

* working segmentation load

* segmentation panel

* add frame of reference uid to monai segs

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* make probe tool grab color for selected segment

* parse centroids and add to segmentation

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* initial commit for toolbox

* fix merge

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------


* feat(labels): add initial labels from the server as starting point (#1546)

* working segmentation load

* segmentation panel

* add frame of reference uid to monai segs

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* make probe tool grab color for selected segment

* parse centroids and add to segmentation

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* initial commit for toolbox

* fix merge

* add initial segments from the server

* fix after pull

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------


* Add batch infer


* Change background colors and fix click edition


* change ohif image


* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Proxy Headers


---------

+++ monailabel/datastore/utils/convert.py
@@ -60,7 +60,7 @@ def dicom_to_nifti(series_dir, is_seg=False):
     return output_file
 
 
-def binary_to_image(reference_image, label, dtype=np.uint16, file_ext=".nii.gz"):
+def binary_to_image(reference_image, label, dtype=np.uint8, file_ext=".nii.gz"):
     start = time.time()
 
     image_np, meta_dict = LoadImage(image_only=False)(reference_image)
